+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 
 
-## 0.13.0 — unreleased
+## 0.13.0 — 2026-09-08
 
 ### Added